Today, I hunted bugs (yum)! The first bug regarded a strange behavior I noticed while making our 50% video: after joining a debate, a user's main screen would show duplicates of all new debates. Additionally, I fixed an error in our binary search algorithm which was incorrectly deleting debates. I was a bit frustrated with how long these tasks took, but that's the life of a bug hunter!
Areas I am struggling in: I think it is going pretty good right now. Whenever you involve multiple devices in data access, there are a lot of issues with permissions and security. Luckily, Firebase handles a lot of this for us, but it can still get tricky at times. Today was a pretty light day in implementing new features, but I tend to find that a lot of strange edge cases can come up when trying to change how database access works.